requirements.txt
flask
groq
ddgs
python-dotenv
gunicorn